[precompile] Add CompilePackage to serialize dynamo states.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/155118

Adding a per torch.compile() object CompilePackage which tracks dynamo artifact. CompilePackage is considered a low level component and should not be directly exposed to end users. It has the following interface:

1. `CompilePackage.__init__()` which optionally takes previously serialized dynamo states.
     a. when `dynamo` argument is None, it will contruct a brand new CompilePackage object.
     b. when `dynamo` argument is not None, it will load a pre-compiled dynamo state.
2. `package.save()` which dumps the dynamo states to a _DynamoCacheEntry type.
3. `package.install(backends)` which will handle all the side-effectful global scope updates with compiled functions and resume functions.

This diff focus on making the low level mechanism for precompile. It will be left to upper level interface to use these API to build more user-facing frontend.
ghstack-source-id: 283346c75b067509c67803c4e922eaee190695d6
@exported-using-ghexport

Differential Revision: [D75956538](https://our.internmc.facebook.com/intern/diff/D75956538/)