Add AOTDispatcher config to set backward autocast behavior
This PR adds a new config `backward_pass_autocast`, to set the backward autocast
behavior. It does not change the existing behavior.

The reason why we need this is that torch.compile acquires a forward and
backward graph at the time of the forward pass. This means that
implemented naively, if there are any context managers active outside
the call to torch.compile, the backward graph will also get the
behaviors from those context managers. This PR gives users a way to
tweak the autocast behavior of the backward pass.

Please see torch._functorch.config for the options to the
`backward_pass_autocast` config.

ghstack-source-id: 94252a07de368c6b3d666a44b4a94a7f92bf4245
Pull Request resolved: https://github.com/pytorch/pytorch/pull/156356