Add ability to tell when a non-trivial autocast kernel has been triggered
This PR adds an API to tell if an autocast kernel has done a casting
(that is, the autocast kernel is not a no-op). The implementation is
that we hook into the core function that gets called on all autocast
kernels; when this function does a non-trivial tensor cast we set
a thread-local flag.

Test Plan:
- new tests

ghstack-source-id: 7b6c3403ec4b95f245b0f1f4a04bdbd4c3baecf3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/156739