Add torch._functorch.config.on_backward_autocast_mismatch
If we are able to detect that (1) autocast happened during the backward
pass and (2) the autocast state is different between what we assumed
while compiling the backward pass and runtime of the backward,
then we raise an error, warn, or do nothing depending on this config.

This is "best-effort": if the user specified to us to assume that no
autocast happens during the backward pass, then we are not able to tell
"if autocast happened during the backward pass", so the checks will
never fire.

I also added a doc page around this.

Test Plan:
- new tests

ghstack-source-id: b9501cc6c89f6bcb0595d44db4061164fb21727a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/156780