Move logging into inner method for reorder pass
The reason for inner/outer method is to keep the outer method conforming
to the typedef for a comms graph pass which returns one obj, while
allowing unit tests to call the inner method that returns more metadata
useful for testing the pass.  The logs should be in the inner part, so
they are functional also during unit testing.

ghstack-source-id: 58fac166d95829aeec0c2c3537af0f251817c945
Pull Request resolved: https://github.com/pytorch/pytorch/pull/156879