[AOTI] Add device guard when launching autotune kernels
Summary: Fix https://github.com/pytorch/pytorch/issues/157737. When launching Triton kernels in the autotune block, we need to consider the fact that the model may not always be on device 0. The reason this was not caught on CI is because test_on_gpu_device1 requires multi_gpu and was not run on a multi_gpu instance. Added test_on_gpu_device1 and other similar multi_gpu tests back.

ghstack-source-id: 2954a32fdff98c650a159c3af8a968534aa16e8d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/158034