ci: Switch lintrunner-noclang to use linter image
This changes the image the lintrunner jobs utilizes to be the base linter image
instead of the CUDA image. This is done to reduce the image size and speed up the
build time.

This was switched in https://github.com/pytorch/pytorch/pull/110502 when
clang used to run in the lintrunner jobs but it is now split out so we can
use the default image for non-clang jobs.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: 537c17c5f7aa3870f1d24477dc50d939134ae886
Pull-Request: https://github.com/pytorch/pytorch/pull/158261
Signed-off-by: Eli Uriegas <eliuriegas@meta.com>