ci: Only run lint on changed files for PRs
This more closely mirrors the type of behavior that users
expect when running lint locally on their PRs.

This also leaves the default behavior as a fallback for when
you're not running on a pull request.

Since lint runs on the `pull_request` event I'm not really worried about
any type of ciflow shenanigans in this.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: 8019bd212cf6ed67f2d5508a67da2d55286a1c83
Pull-Request: https://github.com/pytorch/pytorch/pull/158516
Signed-off-by: Eli Uriegas <eliuriegas@meta.com>