[benchmarks] Add scalar loss as model output when training
Add a hook to benchmark model forward passes that calculates a scalar loss as the first output when training (and detaches all other outputs).  This is a requirement to use joint graph export (experimental), but can be done without loss of generality.

Additionally ensures that Dynamo traces through the loss calculation function (not done previously), which reduces the number of graph breaks in models when training.

ghstack-source-id: 937688a50a3f080bede8e241fcc4cfae8a9450dd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/158074